requirements.txt
streamlit
requests
fastapi
uvicorn
google-genai
pydantic